requirements.txt
fastapi==0.88.0
uvicorn[standard]==0.20.0
sqlalchemy==1.4.41
pydantic==1.10.2
python-multipart==0.0.6
requests==2.28.1
beautifulsoup4==4.11.1
python-dotenv==0.21.0
lxml==4.9.1
html5lib==1.1
openai==0.28.1 